Refactor segment picker map widgets
